removes duplicate rows from metadata.tsv
B123 is printed twice, no. 123a with German lyrics (Heilig, heilig) and no. 123b with Latin lyrics (Sanctus, sanctus). The recording identifiers for the first one were kept because it corresponds to the lyrics in the MuseScore file.
For no. 361, the same recording is featured twice on each of the Brilliant Classics releases, every time on two different CDs, as can be seen on https://musicbrainz.org/recording/543f4654-dcdd-40bf-aa8f-699cc7c13071